backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ORM & Database (同步模式，不需要 greenlet)
sqlalchemy==2.0.36
aiosqlite==0.21.0
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI
openai==1.58.1

# Data Processing
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0

# Validation & Config
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Utilities
httpx==0.28.1
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
